requirements.txt
gym==0.26.2
matplotlib==3.10.1
networkx==3.2.1
numpy==2.2.5
pygame==2.6.1
stable_baselines3==2.5.0
torch==2.8.0
tqdm==4.67.1
torch_geometric==2.6.1
torchinfo==1.8.0 
tabulate==0.9.0
ipykernel==6.29.5 
seaborn==0.13.2 
scikit-learn==1.8.0
torchrl==0.11.1
tensordict==0.11.0
wandb==0.25.0
matplotlib-label-lines==0.8.1
